[zero] fix legacy zero import path
