+++ colossalai/zero/legacy/shard_utils/base_shard_strategy.py
@@ -3,7 +3,7 @@ from typing import List, Optional
 
 import torch.distributed as dist
 
-from colossalai.zero.sharded_param.sharded_tensor import ShardedTensor
+from colossalai.zero.legacy.sharded_param.sharded_tensor import ShardedTensor
 
 
 class BaseShardStrategy(ABC):
