+++ colossalai/zero/legacy/shard_utils/tensor_shard_strategy.py
@@ -5,9 +5,9 @@ import torch.distributed as dist
 
 from colossalai.gemini.tensor_utils import colo_model_data_tensor_move_inline
 from colossalai.utils import get_current_device
-from colossalai.zero.shard_utils import BaseShardStrategy
-from colossalai.zero.shard_utils.commons import get_shard
-from colossalai.zero.sharded_param.sharded_tensor import ShardedTensor
+from colossalai.zero.legacy.shard_utils import BaseShardStrategy
+from colossalai.zero.legacy.shard_utils.commons import get_shard
+from colossalai.zero.legacy.sharded_param.sharded_tensor import ShardedTensor
 
 
 class TensorShardStrategy(BaseShardStrategy):
